Treat warnings as errors by default
git-svn-id: https://svn.atnf.csiro.au/difx@1800 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ master_tags/DiFX-1.5.2/applications/vex2difx/src/vexload.cpp
@@ -335,7 +335,7 @@ static VexInterval adjustTimeRange(map<string, double> &antStart, map<string, do
 		exit(0);
 	}
 
-	if(antStart.size() < minSubarraySize)
+	if(static_cast<int>(antStart.size()) < minSubarraySize)
 	{
 		// Return an acausal interval
 		return VexInterval(1, 0);
